backend/backend/requirements.txt
djangorestframework==3.13
djangorestframework-simplejwt==4.8
django-cors-headers==3.9
